requirements.txt
fastapi
uvicorn[standard]
anthropic
python-docx
python-dotenv
python-multipart
pyobjc-framework-Cocoa
pyobjc-framework-WebKit
